MIDI/editor development; bugfixes
IRON NERD: Day 2
If ever again I should deign to place a supposed boolean value in
quotes, bap me.
More development on MIDI handling and the editor. The MIDI is a bit more
of a Puredata nest than I'd prefer; but this is to compensate for
Puredata's silly handling of MIDI in/out. May replace its crisscross of
wires with a tiny Lua file eventually, but it works okay for now.