new feature: edit phrase values
Now it is possible to edit the byte values of notes and commands that
are already in a phrase's sequence, individually or en masse. Note-byte
2 or 3 is shifted by the global velocity modifier, or the global
velocity modifier * -1, depending on the key command.

Amazed, honestly, that I hadn't yet added any sort of feature like this.
I was composing a song, and I thought, "You know what sort of feature
would be useful here? ... Hey, now wait a second."